Update README.md
The goal of this lab is to enhance your understanding of cache coherence. In the lectures you learned the fundamental concepts of coherence protocols. However, even simple three-state protocols, such as MSI (Modified-Shared-Invalid), are more complex in real life implementations. This lab will introduce you to these more complex aspects of coherence protocols such as transient states. This lab is quite challenging. As always, you are encouraged to start early.

At the end of this lab:

You will have a thorough understanding of how coherence protocols are implemented and why we need transient states.

You will have implemented an MSI coherence protocol in the SLICC domain-specific language in the gem5 simulator and have validated its correctness.

All work on this assignment should be done in groups of two.

The remaining of the handout is organized as follows: Section 2 provides an overview of co-herence and introduces you to transient states and other coherence implementation challenges. Section 3 introduces you to the gem5 simulator and the SLICC domain specific language via a walk-through example. Section 4 defines the problem you are asked to solve, while Section 5 provides useful guidelines for code development and debugging. Section 6 specifies the prelab preparation you should complete. Sections 7 and 8 describe the assignment deliverables and a high-level overview of the marking scheme respectively. Finally, the Appendix provides a handy reference of all the provided SLICC actions.